Package and send in queries to the air
We are now supporting sending in queries from the
air to the cloak. Update queries are sent on save,
and batch queries on request.

We have added an authenticity token that is passed
to the cloak along with the query. It is then later
returned to the air along with the results.
Results that don't contain a valid authenticity token
are dropped.