Remove extra events and only show existing events
If there are more than one event for a machine-file version
combo, then the old ones are removed.
If an event is reposted from the same machine for the same
version, then we update the existing event.

Additionally we now only show events that have been
actually generated for a particular file.