Removes percentiles from the system
Percentiles are not used in the cloak anymore.
Instead they are reported as traditional properties
with ranges assigned.

The percentiles were causing all sorts of trouble
in our system since they relied on the hstore table
type in postgresql. This table type is not installed
by default in all PostgreSQL installations, making it
a pain to test our application on systems with a
standard PostgreSQL installation.

The added migration conditionally removes the tables
if it is run on a legacy system where the table existed,
otherwise it basically is a no-op.

This commit also removes the initial migrations that
would create the tables in the first place, meaning if
the full migrations are run on a new system, the percentile
tables aren't created in the first place.