Remove schema from repo
The schema is recreated when you are migrating your db.
It also differs slightly depending on the database you are
using (sqlite3 in development and PostgreSQL in production).
We don't want these changes continously changing the checked
in state. Furthermore it is bad practise checking in code
that is automatically created.