Pry dep breaks production
By mistake pry was included in one of the
lib classes. This was once useful for debugging,
but not at all required in production.

In reality the app breaks in production, since
the production bundle doesn't include pry at all.

This commit also brings the site up to the
lastest message formats