Notify test server when cloak ready for testing
This is a rather complex and large commit.

The following actions take place.

When a new commit is added to the system, a
build is created for it. As soon as the build
has been successfully created, a cluster will
be created for the particular build.
In this incarnation, the cluster is a VM cluster.

Once the cluster creation has completed, the
test server is called with a list of the names
of the machines that are part of this particular
test.