Documented setup process for new servers
Now the install process needed to setup
a server has been taken out of the deployment
script cycle. It is instead documented in
the readme, and should be done as a separate
step.

This simplifies the capistrano deployment
scripts significantly.

We also update the deployment script
to use air1 as the target server.