Add crontab for running nightly version_test
Will run a nightly version_test with the latest
versions from develop.

It will completely ignore if there has already
been a test run for the exact configuration.

This deployment is rather more complex.
It includes a new gem, so it requires a proxy
setup and a double push.
I would like to try to do this myself, just to
finesse the approach.

Closes #198