Add /api/queries/:id/latest_result_id
Returns a ResultsProto containing only the latest
results id. This accounts for the fact that the
results as received by the web might not be in the
same order as the ones sent by cloak-core.
Through this API we always receive the latest
result id generated by cloak-core as known by
the web.