Send notification email on cluster change
Whenever a cluster changes
for better or for worse,
a notification email is sent out.

The email is currently sent out to all
devs on everyone-dev.

Depending on the number of emails,
we might want to reduce it somewhat
going forward, or limit it to the
person responsible for a cluster.

Unfortunately we do not have the
notion of a responsible person at the
moment (and establishing it is not
an immediate neccessity), so blasting
out an email to everyone is a good
place to start.

Closes #277