Fail with sensible error if missing github token
We require a github oauth token on the
server in order to read out meta data
about individual repositories, and create
builds from the latest commits on a particular
branch.

If the token is missing on a production machine
we need to complain loudly.

Most interactions with github should be stubbed
out when running in local mode, so for local
testing the token should not be needed.