Remove parts related to windows client
We previously had a windows client that
would collect user data and upload it
to a cloak.

The windows client was self-updating
and used the web as a source for new binaries.
These binaries were automatically tested
in a staging machine setup before they
would be allowed onto client machines.

None of this is needed anymore.