Breaking change: API for latest result id
Previously the cloak would include a
result id with a result, giving results
a logical ordering. This only worked
as all tasks were fed through the
global cache, which is no longer the case.
We now assume that for all intents and
purposes we will receive results in order.
The API call to get the latest result
now returns the internal ID of the
last result to be received by the web.

This API will be deprecated for a streaming
API at a later point in time.