Allow analysts to manage users themselves
This commit adds a user administration section
for analysts, where they can see their own
users.

Aircloak employees on the other hand get to
see a list of all users, and can also create
new users, give them privileges, and assign
them to particular analysts.

We additionally remove the ability for users
who are not signed in to register, since that
has so far made no sense, and we are not
offering access on a SaaS basis.

Also make it possible to edit user accounts.