showing of lists of results is moved to tasks-view
We move the view to see a list of the latest and all results to
the tasks-view from the results-view.  As we require a task-id
instead of a result-id, it doesn't make sense to use the path
/results/:task_id and thus the show routine of the
results-controller.  Furthermore we show the latest 3 results by
default (all results is provided optionally).