Can add tasks despite having deleted tables
Previously, if an analyst had a table
that had been deleted, the website would
crash when he attempted to add a new task.

The reason was that the tasks form listed
all tables associated with the analyst (including
the deleted ones which do not have any table spec),
in order to populate the menu for prefetching data.

This commit adds a method on the analyst to
only return tables which have not been deleted.