Existing key's don't have PEM version
To avoid confusion for key's that were
created before we added PEM support,
we hide the PEM download button for
legacy keys.

This prevents someone from downloading an
empty PEM key and failing when using it.

We cannot retroactively generate PEM
keys for existing keys as we don't know
the password the user used when creating
the keys.