Make travis callback a noop
Travis sends test pass notifications to the web
which in turn caused us to register a version
for inclusion in a build.

With the previous set of commits, we have moved
away from this model, and now instead create builds
based on branches.

In this commit the travis callback becomes a noop.
Once we have removed the callbacks from the corresponding
repos .travis.yml file, we should remove the controller
alltogether.
I have created the corresponding issue for that:
https://github.com/Aircloak/web/issues/409