Add basic help guides test
This is a rather weak test. Some of the help guides
have branch points, which in this case do not get
tested at all.

What this tests assures is that most of the glaringly
obvious test mistakes that have taken place so far
will be caught before the tests hit production.