Render results client-side
We store the results in a hashmap (transferred as JSON to the browser)
and we render them client-side. This is needed to be able to disply the
results in real-time by listening for Airpub notifications.
This also required some changes to the layout as the previous one didn't make
as much sense with the new design.