Show deployable entity build status
In the previous version, the build log was
only shown _while_ a build was ongoing,
but at that time there wouldn't be any
build log to show, so nothing would
ever be shown.

The current deployment has the fix
already applied, in order to verify
that it was in essence the right fix.

Therefore, no need to deploy.

Closes #439