Fix test issue
I was manually inserting N records into the database,
one of which was tied to a dynamically generated ID and
the others not.

There is the potential for overlap between the static
values and the dynamically generated values, which then
cause the test to fail.

This commit fixes this issue by making all values take
the dynamic one into account.