moving to a new design for repeated answers which allows auto-resolving
The new design includes that we have a database of task-codes which are
marked trustworthy (or not).  If all task codes of a repeated answers
report are marked trustworthy, we assume that we can ignore the report
and mark it automatically as resolved.

Marking a report on the other hand will not mark all the codes
trustworthy (e.g. there is dangerous code involved and we want and have
taken measures to stop the analyst using that task).  We want to mark
the report resolved to indicate that someone already reviewed it and
has taken corresponding countermeasures.