sanitizing cluster messages to manny-air
This commit adds the following changes:

  - machine lists pushed to manny-air are always sorted

  - we sort based on the machine name (to have a deterministic
    sorting which is not based on arbitrary chosen IDs in web's
    database or something similar)

  - we ensure that each cloak is in the list exactly once

With the previous commit it should never happen that a cloak is
multiple times in a cluster (except if a cluster is created
manually through a ruby command line).  This commit ensures that
even if the database is in such an unclean state, that we will
never send out messages where a cloak is listed in a cluster
multiple times.