Fix message race-condition on latest results page
The problem was that the rails view took so long to render
(~300ms on my machine, plus redirect and then a subsequent
page render that took about 400ms), that the cloak would
already have responded with the result for fast tasks.

Additionally I respond with json in the tasks results
controller, this cuts down the rendering time to
a total of 16ms on my machine, and removes the need
for an additional page load.

A bug was also fixed where task errors would cause
the rendering to fail.

Closes #489