Split destroy task and delete results
There was confusion in the interface where
the same button allowed for removing results
and destroying the task, and it was modal depending
on whether there were results for a task or not.

The "destroy task" buton now only destroys tasks,
and never has any other function.

On the results pages I have now added a new
button for removing all results. It has its
own controller action, and is clearly separate.

The method name `efficient_delete` in the task
which only efficiently deletes results, rather
than the task _and_ it's results was misnamed.
It is now called `efficiently_delete_results`
to avoid confusion.

Closes #510