Analyst deletion should not remove admin users
In most cases, when removing an analyst, one
also wants to remove the associated users.
This is NOT the case when one of those users
is an Aircloak employee who happens to be
impersonating the user.

This commit manually removes all the users
as long as they are not an administrator.

Closes #608