Add GET resource action for individual result
This end-point will be used when the latest_results
page detects that a new answer has been delivered by
airpub which requires post-processing.
This post-processing is done server side (the logic
doesn't exist on the client), so the client then
fetches the particular result from the backend
using this resource action.