Produce comrpessed graphable data
Rather than producing a significant amount
of post-processed buckets that explode the
analyst screen, we now produce one combined
bucket containing all the data needed to
perform the rendering.

This cleans up the display, and paves the
way for the ability to render multiple
distinct graphs on the same page.