removing CloakHelpers and only use cloaks belonging to the cluster
The code of CloakHelpers (which was a copy&paste of code from
app/models/task.rb) was basically unused.  The only information
used was if there is a cloak available for a cluster.

This commit now makes cluster.random_cloak_ip to return only an
IP of a cloak that belongs to the cluster and nil if there is no
such cloak.

Users of this method are modified to handle nil and perform
reasonable error handling in that case.  The functions defined
in CloakHelpers are no longer used and thus the module is removed.

Closes #654