Move CSV export to Erlang app
The current CSV export implementation is rather
inefficient as it does two full passes through the
data being exported. The reason for this design
choice is that this way we never need to hold the
full data in memory at once. As a result we can
export very large reports without blowing the
memory of the server.

It works as follows:

- 1st pass: go through all properties to report
  building up a set of all the available columns
  that we want to be part of the CSV
- 2nd pass: construct rows with empty columns
  where needed, and stream them to the client
  row by row

We also introduce scaffolding on the erlang side
to authenticate the user through a rails API.