Add proxy resource interacting with rails
Introduces a proxy request which handles all
requests under /backend/* not explicitly handled
by another resource.

The resource forwards the request to the rails
backend under the same path (minus /backend/).

It expects an JSON response with a description
of what module to invoke, and with what arguments.
The RPC handler is then expected to fully handle
the response to the client.