Move processing out of DB loop
Each request has two processes working:

- process getting data from db
- process waiting for results to push to the client

Previously we were doing all the processing
on the DB loop itself, having the process
sending data to the client waiting idle.
Now the DB thread sends the data as it receives
it straight on to the other process where the
processing takes time. This way, the DB thread
can move on to the next row in the meantime.

This has some benefits:

- faster downloads for the client
- less data moved between DB and sending process
- we can release the DB connection faster
- simpler code

Possible drawbacks:

- if the sending thread cannot keep up,
  it might end up flooded with messages
  and cause memory spikes.