Add integration test in backend
The integration test creates a cluster with a build of the latest
state of the develop branches. It then creates a table on the cluster,
uploads data to it and runs a query against it.

It evaluates that the whole suite of our infrastructure works.
Whether we can create builds and clusters. If our web infrastructure
works, and to what extent airpub is up and running.

It evaluates if our clusters are able to accept a large amount of data,
and run expensive queries.

The tests heavily rely on the rails app and its integration test API
to perform the individual steps of the tests.