Multiple integration test fixes
This is a series of fixes combined into a single commit.
It adds more verbose and explicit logging during the test,
so we get visibility into what state the test is currently in.

The test now also has the option of aborting the data upload
prematurely should there be a problem.

Additionally we no longer wait after a cluster has been
created. When manny-air reports that a cluster is ready,
it has in fact been migrated and all is good, so we can
go ahead and use it! God bless manny-air for its hard
work.