Move alteration logging to after create
Currently we were logging an alteration
irrespective of the cluster was created or not.
This isn't strictly speaking critical per se,
as the alteration would get cleaned up if the
creation didn't take place, but what was
problematic was that the alteration description
relied on parameters which weren't always there.
Now logging the alteration after the creation
has taken place we know for a fact that
the cluster will have a name, and also a build.