rearrange inter-service dependencies
Backend and frontend now depend on router, which means that stopping a router will stop both services. This is needed since each service need its router to talk to the rest of the world. For example, if the router is dead, backend requests to other services will fail, so it's better to stop a partially working service as well.