improve caching of temp initialization
Instead of generating temp file locally and then copying, we add a pseudoinstruction which is replaced before the build. This approach allows us to cache the contents of the initialization file via docker layers.

Otherwise, accidentally touching the temporary file would cause all images to be rebuilt and bump the version, which is not very nice, not only because of build time, but also because it would cause needless restarts.