remove useradd hack
Previously, we used this to ensure proper permissions on a Linux system when building the backend release. Without this hack, files built in the builder container didn't have proper permissions, and the host user couldn't access them.

However, we're currently building the complete release during the build process, so there's no need for this hackery anymore.