Remove uniqueness constraint of task name
This has been bugging me for quite a while.
Since we allow tasks to be deleted and then
undeleted, we could not reuse the names of
deleted tasks! In case they would be undeleted
in the future, that would have caused a name clash.

I have now removed the uniqueness constraint on
task names, and it is then the user's responsibility
to provide names that make semantic meaning to them.

The names are not used in APIs anyway, so the uniqueness
constraint was kind of arbitrary.