Fix integration test for local environment
The integration test had a race condition whereby
the cluster creation would never succeed.
Since in local mode we immediately mark a cluster
as created, the postgres notification was sent
before the test server would start listening for it.

This fix makes the integration test immediately
continue the test if it is not running in global
mode where a cluster actually has to be setup.