Turn integration test results into a bar graph
A stacked bar graph makes it easier to spot if the
test as a whole has gotten better or slower.
Additionally it can be flipped into a grouped
graph to make it possible to see if individual
phases have improved or gotten worse.

An additional fix was made that turns undefined
values into 0 rather than `undefined`.