Removes task API from cloak API doc
The cloak task API is not encouraged.
When users query the cloaks through the
web, we retain a whole lot more control.
For example can we augment the queries
with task libraries, and we can in
other ways rewrite the queries to
make them better, as well as doing
post-processing of the results!

All queries in the future are likely
to go through the air in some form anyway.
For example consider non-TM queries
which for example approximate some
SQL subset. These would certainly
go to the air, to be rewritten into
a format understood by the cloak.
They might very well also be de-composed
into several queries.

This commit is a step in that direction.