Attempts to fix the csv_rpc test failure
I hope this can fix the problem, but I am rather
unsure. It is not a fix I have been able to validate,
as I have never been able to reproduce the csv_rpc
module test failure locally, and it only periodically
occurs on Travis.

The following stackoverflow post mentions that httpc
can have problems when the `Content-Length` header is not set:
http://stackoverflow.com/questions/23388829/erlang-error-received-unexpected-tcp-data-on-port
They suggest using a `relaxed` option in httpc if
the server side is misbehaving. The server side ISN'T
misbehaving though. We are doing a streaming response
with the `Transfer-Encoding: chunked` set, which
means that we cannot set the `Content-Length`-header
(we don't know the length of the response, and the
spec states that if `Transfer-Encoding` is set, then
any `Content-Length` header provided should be ignored:
https://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html
See 4.4.3).
I have validated that webmachine sets the `Transfer-Encoding`
header correctly: it does.

This commit replaces the usage of `httpc` with `hackney`,
hoping that it _magically_ solves the issue.