Add basic user scaffold
This effectively is a handwritten copy of
the standard scaffold. Only difference is
that I now understand what is going on.

Inspired by the user authentication system
described in Programming Phoenix and Aeacus.

What currently is implemented is password hashing
and validation with Comeonin. A user controller
exists that allows the creation of users with
a name, email address and password.
The email address is the login.

A session controller exists that allows a user
to validate a login. No session is written
when the user is in fact logged in.
Logging out is not yet supported either.