Adds users link in menu
This is a bit of a controversial change.
In order to get the buttons to align correctly,
I changed the sign out operation to become a `get`
rather than a `post`. This was done such that
I wouldn't have to use a button for the signout link,
which wouldn't be styled correctly...

I think this might be OK for now, in the initial
scaffold?