Enable shortcuts also when editor is not active
This is something that has been annoying me for quite a while.
We had bound the Ctrl-R and Ctrl-S keys inside the codemirror
editor itself, which meant they did not work when the editor
was not actively in use. This commit double binds the keys, so
we are sure to capture the user intent.