Simplify params construction and validation
There were too many error conditions and edge cases
in the previous design. I have now simplified it
into something that is somewhat easier to read.

I sacrificed one bit of functionality in the process,
and that was the step of validating that the tables
parameters was indeed a list. In the end that
really is something that should live in the
the task, and not in the controller here.