Fixes click target on tables checkboxes
This has pained me for quite some time :)
The hit target on the checkboxes for tables
in the settings menu was just the checkbox,
and not the checkbox label itself. This made
using the checkboxes a pain.

This commit changes the description to be
within a label, with makes the label itself
clickable too.