Message task_coordinator about job_runner death
This is the simplest way I can think of that does not involve changing
how job_runner <-> task_coordinator communication is done. It's not
ideal, but I've been banging my head against this wall for some time
now, and anyway this is only for failing instantly when things are
already in a pretty bad shape.