make count(*) work with dsproxy
Current implementation of count(*) is a hacky one, since it selects a dummy column called "*" with a value '*' in the place of the count expression. The column name presents some problems for dsproxy.

This commit keeps the original hack (because proper solution requires deeper changes), but makes some minor modifications:

- NULL is selected instead (since we don't care about the value anyway)
- the dummy column has a more descriptive name (ac_ignore)