store cloak info through gproc
Previously we used ETS table for this. That presents a problem when data source becomes large, since ETS doesn't handle larger values well. The problem is amplified since we're storing all tables and all columns. Moreover, we're encoded the data to binary erlang, and then to base64, which added even more to the total size.

With this commit, we're keeping cloak info in the ETS table on the same node where the main channel process is running. This will allow us to efficiently fetch the data, avoid any extra serialization, and relieve the pressure from etcd.

Note that we're still using etcd to store the channel pid. This keeps thing consistent. A node will first fetch the channel pid from etcd, and then the rest of the data by issuing rpc call to that node.