Update compiler to allow cross-joins
As a result the validation of whether columns
exist in tables or not has been moved out of
the compile columns function, and further up.
All column validation is now done while attempting
to qualify the columns, which in an of itself
is somewhat sensible.

The rearrangement also fixes one query_test test,
where a non-existant column was selected, yet another
error was expected.