Select a non-null user-id
OUTER JOINs might result in one of the selected
UID columns to be null. This effectively collapses
all the data across all the null-values into a single
"null"-user, which completely destroys any and all
analysis done across the users.

This pull enables the compiler to deal with data
and uid columns separately. As a result the builder
can ensure that at least one non-null uid column
makes it into the final result. In the case of
standard SQL servers this is achieved using coalesce.