improve parsing of join expressions
The problem of previous approach is that it relies on `option` to chain join expressions. This means that we can't report an error in a join expression.

The previous solution solved this some hackery, generating `{:join, :error, ...}`. Unfortunately, that approach is limited and painful. A problem is that since we now support subqueries, we can't report an error in a subquery which is a part of a join expression.

The solution is to use forward parsin. As soon as we see construct such as `inner join`, we know that the next part must be a valid inner join `rhs`. Arbitrary chaining can be done with `switch` and `lazy` instead of `option`.