compute all required db columns in the compiler
The previous version maintained `db_id_columns` and `db_data_columns`. That approach complicates the work of the data driver which needs to do special post-processing, depending on how much elements are in `db_id_columns`.

Moreover, it makes it harder to support subqueries, where id column is not auto-selected.

This version introduces an explicit concept of a `db_function`, which allows the compiler to provide complex select items. In this case, the compiler generates an abstract `coalesce` db_function call which needs to be interpreted by the driver.

That makes the approach more explicit. The compiler produces a query which fully explains what needs to be done, while the driver decides how will it do it.