Refactor latest_data_source
It had convoluted logic that looked
for the last used data source that is
currently online.

This has now been simplified to instead
take the last query made by the user,
and then go to the data source related
to that query, whether or not that
data source is online.