Fix DataSourceManager crashing the cluster
When starting with a global name a process cannot be normally started as
part of the supervision tree. This moves the global registration to the
init callback instead. It also adds monitoring of the currently
globally registered DataSourceManager so that another one can try to
register if the global one goes down.