Mark proper as a package that's not shipped
Proper is licensed under the GPL, so we cannot include it in our
distribution. However, we use it only for tests, so we can just ignore
it. I opted to mark it here rather than try to detect which packages are
test-only simply, because the latter option seems rather complicated,
especially with nested dependencies.