Use a dedicated server for DataSourceManager tests
The tests have been flaky. It seems to be related to the Ecto.Sandbox
and connection checkouts surviving through a given test, but then
failing when the owner (test process) of the given connection is no
longer around. This way the only processes using a connection are the
test and the server spun up for that particular test.