Add flow and refactor for smaller API
The GraphData class is now turned into a normal
JavaScript object such that the API could be limited.
Otherwise the number of functions that needed to be
typed was never ending, and I needed to get into
the specifics of typing the traces, which really
should be considered opaque to the caller.