Improve chart styling
Moves legends to the top right,
and moves the chart css styling
to a CSS file

Also adds a significant amount of
marging under the graph in order to
show larger labels with their full
values. Unfortunately this results
in there being a large empty white-space
underneath the graph when no label
text is provided. This is a compromise
at best.