Add low count filtering for IN and NOT IN
This adapts the LowCountFilterCondition checker
to allow for two strategies: "drop" and "keep"

The "drop" strategy starts dropping rows as soon
as the low count filter threshold has been passed.

The "keep" strategy immediately passes rows on for
further processing as soon as the low count filter
threshold has been reached.

"drop" wins over "keep", so if two filters match
a row and one of them dictates that the row should
be dropped, then it will.