Add node to .tool-versions and improve .travis.yml
This makes sure that we run with the same version
of Erlang, Elixir, and NodeJS on Travis as we do
locally, and also ensures this remains true without
future needs for changing the travis config.

Previously we had Erlang and Elixir under .tool-versions
control, but not NodeJS. But none of them were used on travis.

The use of asdf requires sourcing of the asdf config
in various places. To simplify this, I extracted out
the travis pre-script and script sections into separate
files where we have more control.