power DataSourceManager with :gproc
I've noticed in a separate PR that a failure of a DB operation during registartion will crash the complete data source manager. Looking at the code, it seems to me that the manager is overly complex.

Since we don't support clustering of airs, we can remove the manager process completely, and thus simplify the code a lot. As an added bonus, a db failure while registering one cloak won't affect others.