Adds licenses for all scrivener dependencies
The current versions of scrivener and scrivener_ecto
have this license explicitly on github. It seems the
version pushed to hex might not have the license included
yet. I think this should be OK.

scrivener_html does not yet have a license.
I have created a pull
(https://github.com/mgwidmann/scrivener_html/pull/41)
with an sample MIT license text. This text (provided
by me) is what is part of this commit).
If he decides not to use the sample text, or to license
the library differently, we'll have to update the license
used.

If he doesn't want to provide a license at all, we'll
have to write his functionality ourselves. It should
be doable.