Adds more test data to local environment
Adds:

- customers
- accounts
- transactions

Where each customer has one or more accounts,
and each account has multiple transactions.

Each customer currently only has accounts in
the same country, but potentially in different banks.

The data is pretty dumb to be honest, but gives something
to work with.

There is currently also a customer_id in the transactions,
to account for the fact that we don't know how to
deal with tables that don't contain a user_id.

The script that generated the accounts and transactions,
can be found here:
https://gist.github.com/sebastian/a2a9badbf56163afb641f6b8e32acd69