stabilise integration tests
It appears that things don't happen as fast on Travis. The default timeout for `assert_receive` is 100ms, and looking at the output, I can tell that server disconnected, but didn't still reconnect. Therefore, in this commit, I increased the timeout for `assert_receive`, and I'm also testing that the server has indeed disconnected.