Partition row-splitter functions
We split from the first row splitter function.
Given the sequence of functions:

  a(b(c(d(column)

where b is a row splitter, we partition it into:

  result_column = b(c(d(column)))
  a(result_column)

where a get's executed on the result of the row splitter.

The row-splitter can itself contain other row splitters.

No row splitting is done at this time.