Bug fix: allowing multiple occurrences of row splitters
Take for example the following query:

  SELECT
    row_splitter(column) as col1,
    row_splitter(column) as col2,
    count(*)
  FROM table
  GROUP BY col1, col2

had `row_splitter(column)` been a non-row-splitting
construct, we would only have needed a single db-entry
for `column`. In the case of row-splitters on the other
hand we are expecting the cross product of the split results
of the functions, and therefore cannot collapse these
columns down to a single DB row instance.

This commit introduces a version of Enum.uniq/1
targeted at row splitters, which account for them
always being distinct.