Adds the capability to validate illegal transformations
Checks top-level selected columns for an occurrence of:
- dangerous math AND
- dangerous discontinuity

Checks inequalities in all WHERE-conditions for
- dangerous math, OR
- dangerous discontinuity

Dangerous math is math where one of the parties involved
in the math has been part of an expression where a constant
was present. For example:

  <column> + 1 -- dangerous
  <column> + <column> -- safe

Dangerous discontinuity is a discontinuous function where
the entities being processed have been through an expression
containing a constant:

  ceil(<column> + 1) -- dangerous
  ceil(<column>) -- safe

Discontinuous string functions are only treated as dangerous
if they are later converted to a number:

  length(btrim(<column>, 'constant')) -- dangerous
  length(btrim(<column>)) -- safe
  btrim(<column>, 'constant') -- safe (not converted to a number

Doesn't yet remove the math feature. This will be done in
a separate pull.

Closes #877