Tweak bound in FixAlign test
There are valid (for the current logic) cases when an input interval
spans less than 30 days, but has endpoints in non-consecutive months.
For example an interval 2000-01-31 - 2000-03-01, but also intervals that
have endpoints mid-day. In that case it might be that the closest
aligned interval is half a year which is slightly more than 6 times the
original in size.