relocate code not specific to db emulation
Aggregator invoked DbEmulator even when no emulation took place. The reason is that functions not related to emulation were included in the `DbEmulator` code. This is confusing and misleading, so I moved these functions to the aggregator module. Perhaps that's also not the best place and we should consider creating another module. However, at this level of my understanding of the code, I'm not sure, so I chose what I consider to be lesser of two evils.