Add a filter in ShrinkAndDrop
ShrinkAndDrop can be order-dependent in some cases if the data has not
been filtered previously. In particular if items at the right boundary
arrive (which they should not, because the ranges we use are exclusive
on that side) some of them will be let through, depending on the order
of arrival. Adding these filters makes sure, that no such items make it
through SaD.