Make HalfBuffer order-independent
See 755d792

While the logic from that commit kind of explains why ShrinkAndDrop
would be order-dependent, the heart of the matter should be solved by
this one. Long story short - the right side of an interval was
improperly excluded in ShrinkAndDrop. Given [0, 0, 0, 5, 5, 5] the
smallest, right-open, 1-2-5-aligned interval that includes all that data
is [0, 10) instead of [0, 5).