Don't double check queries for illegal WHERE-clauses
The compiler compiles each query and sub-query individually.
In the step where WHERE-clauses where checked for illegal
constructs we attempted (and failed to do so, but didn't notice)
to check all subqueries as well (but only checked one level deep).
As a result every single query (but the top-level one) got
typed and checked twice.