Only start doing projections once a threshold is reached
This prevents us from accidentally killing queries
if there is a very large memory spike that is completely
within safe bounds. Say we go from 10GB of free memory
to 5GB of free memory within a very short time frame.
The app could easily thing that we are going to reach 0
if this continues.