Hack around postgres connections lacking type information
If a migration introduces a new type some (?) postgrex connections seem
to have wrong type info, but the thing recovers after a single failure.
This does not work out well for tests though. Starting just air allows
it to migrate in peace and then when the tests are actually ran all
connections have the newest type info from the start.