Remove cloak memory channel
Sends memory readings on the main channel instead.
This is required in order for the cloak service to
be able to associate the reading with the correct
cloak in a straight forward way.

Additionally the cloak memory readings are now
used to directly display the latest value when
loading the activity monitor page.