Make the timeout in Air central socket configurable
In integration tests we're sending usage info every 100 ms. This is
often enough that the first request hits the socket before it even
joined the channel and fails with a timeout. However with the timeout
being 5s it waits a really long time before trying another request
(because the requests are now serialized).