Immediately show pending result
This change makes it so that we are showing
a pending result view even before the server
has accepted the query.

This is very useful when you are sitting on
a slow high latency connection, and aren't
really sure whether or not your "run"-button
click actually registered or not.