Reject queries with non-aggregate functions
This makes the optimiser overly strict, but
this is sufficient for an initial version.

Better safe than sorry. Allowing arbitrary
functions is somewhat more complex, and will be done
later. For example it needs to be taken into account
whether or not the function supports being moved to the
inner select, and if not, whether the data it needs
should be carried out to the outer scope,
or the query optimization altogether forbidden.