Reject queries with WHERE-clauses
This is a short term simplification.
WHERE clauses can usually be moved into
the inner select, but this doesn't always
apply as there are for example complications when
WHERE-clauses are used in conjunction with functions.

This can be improved over time, but is outside
the scope of a first initial implementation.