Make it possible to provide class to navlink helper
The navlink helper now supports a "class" options
argument which will be used in the generated <li>-element.

The helper is furthermore improved to suppress the class
html parameter altogether when there are no classes that
should be applied. This happens when a navlink is not
active, and no additional class was provided for the element.