refactor aggregation and result pipeline
I felt there were too many functions exported, such as bucket_expressions and bucket_expression_index!. The responsibilities were somehow spread all over, so I attempted to consolidate this.

With this commit, the aggregator spits the final result. The post-processing of aggregated buckets (sorting, limiting, ...) is done directly in `Result.new`, so `Result.t` always contains the final result. This also drastically reduced the amount of public functions which we need.

To make this happen, I also had to dig a bit into sorter and the emulator selector.